Backend/requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
peft
python-dotenv
lime
numpy
scikit-learn
